post 통합테스트완성, post 인증 및 허가 추가

+++ src/main/java/com/marklog/blog/web/PostController.java
@@ -1,7 +1,10 @@
 package com.marklog.blog.web;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.marklog.blog.config.auth.dto.UserAuthenticationDto;
+import com.marklog.blog.domain.user.Role;
 import com.marklog.blog.service.PostService;
 import com.marklog.blog.web.dto.PostIdResponseDto;
 import com.marklog.blog.web.dto.PostResponseDto;
@@ -28,31 +33,67 @@ public class PostController {
 
 	@ResponseStatus(value = HttpStatus.CREATED)
 	@PreAuthorize("isAuthenticated()")
-@PostMapping("/post")
-	public PostIdResponseDto save(@RequestBody PostSaveRequestDto requestDto) {
+	@PostMapping("/post")
+	public ResponseEntity<PostIdResponseDto> save(@RequestBody PostSaveRequestDto requestDto) {
 		Long id = postService.save(requestDto);
 		PostIdResponseDto postIdResponseDto = new PostIdResponseDto(id);
-		return postIdResponseDto;
+		return new ResponseEntity<>(postIdResponseDto, HttpStatus.CREATED);
 	}
 
-
 	@GetMapping("/post/{id}")
-	public PostResponseDto findById(@PathVariable Long id) {
-		return postService.findById(id);
+	public ResponseEntity<PostResponseDto> findById(@PathVariable Long id) {
+		try {
+			return new ResponseEntity<>(postService.findById(id), HttpStatus.OK);
+		} catch (IllegalArgumentException e) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@PreAuthorize("isAuthenticated()")
+//	@PostAuthorize("hasRole('ADMIN') or #userId==principal.id")
 	@PutMapping("/post/{id}")
-	public PostIdResponseDto update(@PathVariable Long id, @RequestBody PostUpdateRequestDto requestDto) {
-		Long updatedId = postService.update(id, requestDto);
-		PostIdResponseDto postIdResponseDto = new PostIdResponseDto(updatedId);
-		return postIdResponseDto;
+	public ResponseEntity<PostIdResponseDto> update(@PathVariable Long id, @RequestBody PostUpdateRequestDto requestDto,
+			Authentication authentication) {
+		try {
+			PostResponseDto postResponseDto = postService.findById(id);
+			Long updatedId = postService.update(id, requestDto);
+			PostIdResponseDto postIdResponseDto = new PostIdResponseDto(updatedId);
+			Long userId = postResponseDto.getUserId();
+			Long authUserId = ((UserAuthenticationDto) authentication.getPrincipal()).getId();
+			boolean hasAdmin = authentication.getAuthorities()
+					.contains(new SimpleGrantedAuthority(Role.ADMIN.getKey()));
+
+			if (hasAdmin || userId == authUserId) {
+				return new ResponseEntity<>(postIdResponseDto, HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+			}
+		} catch (IllegalArgumentException e) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+
 	}
 
 	@ResponseStatus(value = HttpStatus.NO_CONTENT)
 	@PreAuthorize("isAuthenticated()")
+//	@PostAuthorize("hasRole('ADMIN') or #userId==#authUserId")
 	@DeleteMapping("/post/{id}")
-	public void userDelete(@PathVariable Long id) {
-		postService.delete(id);
+	public ResponseEntity userDelete(@PathVariable Long id, Authentication authentication) {
+		try {
+			PostResponseDto postResponseDto = postService.findById(id);
+			Long userId = postResponseDto.getUserId();
+			Long authUserId = ((UserAuthenticationDto) authentication.getPrincipal()).getId();
+			boolean hasAdmin = authentication.getAuthorities()
+					.contains(new SimpleGrantedAuthority(Role.ADMIN.getKey()));
+
+			if (hasAdmin || userId == authUserId) {
+				postService.delete(id);
+				return new ResponseEntity(HttpStatus.NO_CONTENT);
+			} else {
+				return new ResponseEntity<PostIdResponseDto>(HttpStatus.FORBIDDEN);
+			}
+		} catch (IllegalArgumentException e) {
+			return new ResponseEntity(HttpStatus.NOT_FOUND);
+		}
 	}
 }
